+++ test/unit/sonarqube_code_coverage.sh
@@ -10,6 +10,8 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mysql_perf test/unit/mysql_perf/convert_dict.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/create_data_config.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/create_header.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/data_out.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/help_message.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/main.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat.py